fix(data): fix datafiller fetch scripts
