+++ packages/code-du-travail-data/dataset/datafiller/README.md
@@ -8,6 +8,9 @@ Pour générer le fichier qui servira à l'indexation, la commande
 $ yarn start
 ```
 
-va générer le fichier `prequalified.json` qui sera ensuite utilisé par [l'API](../code-du-travail-api/routes/search/search.prequalified.js).
+va générer :
+
+- le fichier `prequalified.data.json` qui sera ensuite utilisé par [l'API](../code-du-travail-api/routes/search/search.prequalified.js).
+- le fichier `themes.data.json` qui sera ensuite utilisé par [l'API](../code-du-travail-api/routes/themes/index.js).
 
 Le serveur elastic peut être configuré avec `process.env.ELASTICSEARCH_URL` et utilise `http://127.0.0.1:9200` par défaut.
